Refactor of abstract ImageRaster.
Contains initial implementations for getPixel and setPixel and defines some new optional methods as well as two abstract methods that are implemented by children.
